+++ CHANGES.md
@@ -10,10 +10,12 @@ To be Released.
 * Fixed cursor not updating when buffer height changed.  [[#17]]
 * Fix an error where the system terminal fails to catch a TaskCanceledException.
   [[#19]]
+* Added the feature to parse custom types.  [[#20]]
 
 [#13]: https://github.com/s2quake/commands/pull/13
 [#17]: https://github.com/s2quake/commands/pull/17
 [#19]: https://github.com/s2quake/commands/pull/19
+[#20]: https://github.com/s2quake/commands/pull/20
 
 
 6.0.1